chore: move persist plugin
