--- apps/codeimage/src/core/store/trackContext.ts
@@ -1,15 +1,0 @@
-import {asyncScheduler} from 'rxjs';
-
-export function createTrackContext() {
-  let track = true;
-  return [
-    () => track,
-    (cb: () => void) => {
-      track = false;
-      cb();
-      asyncScheduler.schedule(() => {
-        track = true;
-      });
-    },
-  ] as const;
-}
